requirements.txt
mistune>=3.0
python-docx>=1.1
Pillow>=10.0
requests>=2.31
charset-normalizer>=3.3
